backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv==1.0.1
hume
google-genai
pydantic
aiofiles
